+++ ChangeLog.md
@@ -8,6 +8,13 @@
 * Fuzzy matching support to give better error messages for close version numbers [#504](https://github.com/commercialhaskell/stack/issues/504)
 * --local-bin-path global option. Use to change where binaries get placed on an `stack install` [#342](https://github.com/commercialhaskell/stack/issues/342)
 * Custom snapshots [#111](https://github.com/commercialhaskell/stack/issues/111)
+* --force-dirty flag: Force treating all local packages as having dirty files (useful for cases where stack can't detect a file change)
+* GHC error messages: display file paths as absolute instead of relative for better editor integration
+
+Bug fixes:
+
+* Extensions from the `other-extensions` field no longer enabled by default [#449](https://github.com/commercialhaskell/stack/issues/449)
+* Fix: haddock forces rebuild of empty packages [#452](https://github.com/commercialhaskell/stack/issues/452)
 
 ## 0.1.2.0